Refactor interpreter interface
