+++ python/tvm/relay/graph_runtime/__init__.py
@@ -1,0 +1,2 @@
+from . import codegen
+from . build_module import build
