+++ python/tvm/relay/graph_runtime/codegen.py
@@ -23,116 +23,12 @@ We expose this functionality in compile_to_tvm.
 from __future__ import absolute_import
 import json
 import attr
-from . import ir_pass
-from .op import Op
-from .expr import Var, Function, Call, If, GlobalVar, Constant, Let, Tuple
-from ..build_module import build as tvm_build_module
-from .. contrib import graph_runtime
-from .ir_pass import infer_type
-from .. import cpu
-
-class AbstractExprVisitor(object):
-    """A visitor over Expr in Python."""
-
-    def __init__(self):
-        self.memo_map = {}
-
-    # pylint: disable=no-else-return
-    def visit(self, expr):
-        """Apply the visitor to an expression."""
-        found = self.memo_map.get(expr)
-        if found:
-            return found
-
-        if isinstance(expr, Function):
-            res = self.visit_function(expr)
-        elif isinstance(expr, Call):
-            res = self.visit_call(expr)
-        elif isinstance(expr, Let):
-            res = self.visit_let(expr)
-        elif isinstance(expr, Var):
-            res = self.visit_var(expr)
-        elif isinstance(expr, GlobalVar):
-            res = self.visit_global_var(expr)
-        elif isinstance(expr, If):
-            res = self.visit_if(expr)
-        elif isinstance(expr, Tuple):
-            res = self.visit_tuple(expr)
-        elif isinstance(expr, Constant):
-            res = self.visit_constant(expr)
-        else:
-            raise Exception("warning unhandled case: {0}".format(type(expr)))
-
-        self.memo_map[expr] = res
-        return res
-
-    def visit_function(self, _):
-        raise Exception("Abstract method please implement me.")
-
-    def visit_let(self, _):
-        raise Exception("Abstract method please implement me.")
-
-    def visit_call(self, _):
-        raise Exception("Abstract method please implement me.")
-
-    def visit_var(self, _):
-        raise Exception("Abstract method please implement me.")
-
-    def visit_type(self, typ):
-        return typ
-
-    def visit_if(self, _):
-        raise Exception("Abstract method please implement me.")
-
-    def visit_tuple(self, _):
-        raise Exception("Abstract method please implement me.")
-
-    def visit_constant(self, _):
-        raise Exception("Abstract method please implement me.")
-
-    def visit_global_var(self, _):
-        raise Exception("Abstract method please implement me.")
-
-
-class ExprMutator(AbstractExprVisitor):
-    """A functional visitor over Expr in Python."""
-
-    def visit_function(self, fn):
-        new_body = self.visit(fn.body)
-        return Function(
-            list(fn.params),
-            fn.ret_type, new_body,
-            fn.type_params)
-
-    def visit_let(self, let):
-        new_var = self.visit(let.var)
-        new_val = self.visit(let.value)
-        new_body = self.visit(let.body)
-        return Let(new_var, new_val, new_body)
-
-    def visit_call(self, call):
-        new_fn = self.visit(call.op)
-        new_args = [self.visit(arg) for arg in call.args]
-        return Call(new_fn, new_args, call.attrs)
-
-    def visit_var(self, var):
-        return var
-
-    def visit_global_id(self, global_var):
-        return global_var
-
-    def visit_if(self, ite):
-        return If(
-            self.visit(ite.guard),
-            self.visit(ite.true_b),
-            self.visit(ite.false_b))
-
-    def visit_tuple(self, tup):
-        return Tuple([self.visit(field) for field in tup.fields])
-
-    def visit_constant(self, const):
-        return const
-
+from .. import ir_pass
+from ..op import Op
+from ..expr import Var, Function, Call, If, GlobalVar, Constant, Let, Tuple, ExprMutator
+from ... contrib import graph_runtime
+from ..ir_pass import infer_type
+from ... import cpu
 
 @attr.s
 class NodeRef(object):
@@ -474,78 +370,5 @@ class GraphRuntimeCodegen(ExprMutator):
         return json.dumps(json_dict)
 
 
-def build(env, func, target=None):
-    """
-    Compile a single function to the components needed by the
-    TVM RTS.
-
-    Parameters
-    ----------
-    func: relay.Expr
-        The function to build.
-
-    target: optional str
-        The target platform.
-
-    Returns
-    -------
-    (graph_json, mod, params): tuple of (str, tvm.Module, dict)
-        The outputs of building a Relay function for the TVM runtime.
-
-    """
-    if target is None:
-        target = 'llvm'
-
-    comp = GraphRuntimeCodegen(env)
-    # NB(@jroesch) This creates lowered functions, and generates names for them
-    #
-    # We need these names to emit the correct graph as these are names of the
-    # functions contained in the module.
-    lowered_ops = ir_pass.lower_ops(env, func)
-    mod = tvm_build_module([lf.lowered_func for lf in lowered_ops], target)
-
-    # Therefore the call to compile must come after.
-    comp.codegen(func)
-    graph_json = comp.to_json()
-    return graph_json, mod, None  # params currently isn't supported by API
 
 
-def graph_evaluate(env, func, *args):
-    """
-    Corresponding function to tvm.relay.eval.evaluate.
-
-    This function evaluates a Relay expression on the
-    TVM graph_runtime.
-
-    Parameters
-    ----------
-    env: tvm.relay.Environment
-        The global environment used.
-
-    expr: tvm.relay.Expr
-        The expression to evaluate.
-
-    args: list of tvm.relay.Expr
-        The arguments to apply to the expression, only works
-        if the expression has a function type.
-
-    Returns
-    -------
-    value: tvm.NDArray
-        The output Tensor produced by evaluating the expression.
-    """
-    func = infer_type(func, env)
-    func = ir_pass.fuse_ops(env, func)
-    func = infer_type(func, env)
-    graph_json, mod, params = build(env, func)
-    assert params is None
-    gmodule = graph_runtime.create(graph_json, mod, cpu(0))
-    # Create map of inputs.
-    inputs = {}
-    for i, arg in enumerate(args):
-        inputs[func.params[i].name_hint] = arg
-    # Set the inputs here.
-    gmodule.set_input(**inputs)
-    # Run the module, and fetch the output.
-    gmodule.run()
-    return gmodule.get_output(0)
